+++ android/app/src/main/java/com/whj/screenkit/AdminRecv.kt
@@ -1,5 +1,0 @@
-package com.whj.screenkit
-
-import android.app.admin.DeviceAdminReceiver
-
-class AdminRecv : DeviceAdminReceiver()